refactor: adjust prompts, disabled the ability to call different tools, centralized tool calling
