+++ .vectorcode/config.json
@@ -1,0 +1,5 @@
+{
+  "host": "127.0.0.1",
+  "port": 6131
+}
+
